video_analyst/requirements.txt
# video analysis
phidata
google-generativeai
duckduckgo-search
streamlit
python-dotenv
